Extract SettingsInputComponent creation methods to their own factory object
