+++ src/games/strategy/triplea/settings/ai/AiTab.java
@@ -9,6 +9,7 @@ import javax.swing.JTextField;
 import games.strategy.engine.ClientContext;
 import games.strategy.triplea.settings.IntegerValueRange;
 import games.strategy.triplea.settings.SettingInputComponent;
+import games.strategy.triplea.settings.SettingInputComponentFactory;
 import games.strategy.triplea.settings.SettingsTab;
 
 public class AiTab implements SettingsTab<AiSettings> {
@@ -20,14 +21,14 @@ public class AiTab implements SettingsTab<AiSettings> {
     BiConsumer<AiSettings, String> aiSettingsWriter = (settings,s) -> settings.setAiPauseDuration(s);
 
     inputs = Arrays.asList(
-        SettingInputComponent.buildIntegerText(
+        SettingInputComponentFactory.buildIntegerText(
             new IntegerValueRange(0, 3000, AiSettings.DEFAULT_AI_PAUSE_DURACTION),
             "AI Pause Duration",
             "Time delay (in milliseconds) between AI moves, allows for the AI moves to be watched",
             new JTextField(String.valueOf(aiSettings.getAiPauseDuration()), 5),
             aiSettingsWriter,
             (settings -> String.valueOf(aiSettings.getAiPauseDuration()))),
-        SettingInputComponent.buildYesOrNoRadioButtons("Show Battles Between AIs",
+        SettingInputComponentFactory.buildYesOrNoRadioButtons("Show Battles Between AIs",
             "When set to yes, combats between AI players will be shown in a battle window.",
             aiSettings.showBattlesBetweenAi(),
             ((settings, s) -> settings.setShowBattlesBetweenAi(Boolean.valueOf(s))),
