+++ src/games/strategy/triplea/settings/battle/calc/BattleCalcTab.java
@@ -8,6 +8,7 @@ import javax.swing.JTextField;
 import games.strategy.engine.ClientContext;
 import games.strategy.triplea.settings.IntegerValueRange;
 import games.strategy.triplea.settings.SettingInputComponent;
+import games.strategy.triplea.settings.SettingInputComponentFactory;
 import games.strategy.triplea.settings.SettingsTab;
 
 public class BattleCalcTab implements SettingsTab<BattleCalcSettings> {
@@ -19,14 +20,14 @@ public class BattleCalcTab implements SettingsTab<BattleCalcSettings> {
   public BattleCalcTab(final BattleCalcSettings battleCalcSettings) {
 
     inputs = Arrays.asList(
-        SettingInputComponent.buildIntegerText(
+        SettingInputComponentFactory.buildIntegerText(
             new IntegerValueRange(1, 10000, BattleCalcSettings.DEFAULT_SIMULATION_COUNT_DICE),
             "Default Dice Run Count",
             CALC_DESCRIPTION + " (dice games)",
             new JTextField(String.valueOf(battleCalcSettings.getSimulationCountDice()), 5),
             BattleCalcSettings::setSimulationCountDice,
             (calcSettings) -> String.valueOf(calcSettings.getSimulationCountDice())),
-        SettingInputComponent.buildIntegerText(
+        SettingInputComponentFactory.buildIntegerText(
             new IntegerValueRange(1, 10000, BattleCalcSettings.DEFAULT_SIMULATION_COUNT_LOW_LUCK),
             "Default Low Luck Run Count",
             CALC_DESCRIPTION + " (low luck games)",
