+++ src/games/strategy/triplea/settings/folders/FoldersTab.java
@@ -9,6 +9,7 @@ import javax.swing.JTextField;
 import games.strategy.engine.ClientContext;
 import games.strategy.triplea.settings.InputValidator;
 import games.strategy.triplea.settings.SettingInputComponent;
+import games.strategy.triplea.settings.SettingInputComponentFactory;
 import games.strategy.triplea.settings.SettingsTab;
 
 public class FoldersTab implements SettingsTab<FolderSettings> {
@@ -20,14 +21,14 @@ public class FoldersTab implements SettingsTab<FolderSettings> {
     holderPanel.add(saveGamePathField);
 
     inputs = Arrays.asList(
-        SettingInputComponent.buildTextComponent(
+        SettingInputComponentFactory.buildTextComponent(
             "Save game path",
             "Default save game folder",
             saveGamePathField,
             FolderSettings::getSaveGamePath,
             FolderSettings::setSaveGamePath,
             InputValidator.IS_DIRECTORY),
-        SettingInputComponent.buildTextComponent(
+        SettingInputComponentFactory.buildTextComponent(
             "Map Download Path",
             "Location where maps are downloaded and found",
             new JTextField(settings.getDownloadedMapPath()),
